Move pkg/predicate to controllers/predicate
